backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
PyMuPDF==1.26.4
rapidfuzz==3.14.1
python-docx==1.2.0
reportlab==4.4.4
httpx==0.28.1
pytest==8.2.2
